Reset heads and tails when creating a new supertabular
Otherwise it keeps using the previous ones.

+++ verif/tests/latex/test_gen_price_justification_document.py
@@ -56,7 +56,8 @@ else:
     logging.error('test: '+fname+' ERROR.')
 
 # Remove LaTeX file
-if os.path.exists(thisFile):
-    os.remove(thisFile)
-else:
-    logging.error('ERROR file: '+thisFile+' not found.')
+if (ok):
+    if os.path.exists(thisFile):
+        os.remove(thisFile)
+    else:
+        logging.error('ERROR file: '+thisFile+' not found.')
